Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai>=0.5.0
httpx  